refactor(domain): migrate validators to concrete RuleViolated and ValidationFailed
Replace abstract RuleViolationError/ValidationError throws with concrete
RuleViolated and ValidationFailed in domain validators and composite rule.
