+++ src/forging_blocks/domain/validators/composite_validation_rule.py
@@ -2,7 +2,7 @@
 
 from typing import Any
 
-from forging_blocks.foundation.errors.rule_violation_error import RuleViolationError
+from forging_blocks.foundation.errors.base.rule_violation_error import RuleViolationError
 from forging_blocks.foundation.rules import ValidationRule
