+++ src/forging_blocks/domain/validators/required_validator.py
@@ -2,8 +2,9 @@
 
 from typing import Any
 
+from forging_blocks.foundation.errors.base.rule_violation_error import RuleViolationError
 from forging_blocks.foundation.errors.core import ErrorMessage, ErrorMetadata
-from forging_blocks.foundation.errors.rule_violation_error import RuleViolationError
+from forging_blocks.foundation.errors.rule_violated import RuleViolated
 from forging_blocks.foundation.rules import ValidationRule
 
 
@@ -16,7 +17,7 @@ class RequiredValidator(ValidationRule):
     def validate(self, value: Any) -> list[RuleViolationError]:
         if value is None or value == "":
             return [
-                RuleViolationError(
+                RuleViolated(
                     ErrorMessage(f"'{self._field}' is required."),
                     ErrorMetadata(context={"field": self._field, "code": "required"}),
                 )
